Function for working with linked lists.
The function allows you to initialize the list, print the list
items, add the list items to the back and to the front of list,
free memory after using the list. The function uses the structure
list_t. The function takes arguments: a pointer to the head of
the list and uint32_t.